# swagger.yml
openapi: 3.0.0

info:
  title: OneApp Backend server
  description: This is the first version of OneApp
  version: 0.0.1

servers:
  - url: "{protocol}://localhost:4000/"
    description: LOCALHOST
    variables:
      protocol: 
        enum:
          - http
          - https
        default: http
  - url: http://localhost:5000/
    description: DONT USE

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer  

paths:
  /home:
    get:
      tags:
        - Home
      summary: Go to Home page
      description: Health check
      responses: 
        '200':
          description: OK
          content: 
            application/json:
              schema: 
                type: object
                properties: 
                  responses: 
                    type: string

  /get/insta:
    get:
      tags:
        - Get Data
      summary: Get instagram data
      responses: 
        '200': 
          desciption: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: 
                      type: object
                      properties: 
                        id: 
                          type: integer
                        name: 
                          type: string
                        followers:
                          type: integer
                        follows:
                          type: integer                    
                  date: 
                    type: string
    
  /get/facebook:
    get:
      tags:
        - Get Data
      summary: Get facebook data
      responses: 
        '200': 
          desciption: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: 
                      type: object
                      properties: 
                        id: 
                          type: integer
                        name: 
                          type: string
                        followers:
                          type: integer
                        follows:
                          type: integer                    
                  date: 
                    type: string
  
  /get/linkedin:
    get:
      tags:
        - Get Data
      summary: Get linkedin data
      responses: 
        '200': 
          desciption: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: 
                      type: object
                      properties: 
                        id: 
                          type: integer
                        name: 
                          type: string
                        followers:
                          type: integer
                        follows:
                          type: integer                    
                  date: 
                    type: string
                        
  /get/data/{id}:
    get:
      tags:
        - Get Data
      summary: Get data for debugging purposes
      parameters: 
        - name: id
          in: path
          description: data id
          required: true
          schema:
            type: integer
            format: int64
      responses: 
        '200': 
          descrpition: OK
          content: 
            applications/json:
              schema:
                type: object
                properties:
                  data: 
                    type: string

  /get/data:
    get:
      tags: 
        - Get Data
      summary: Get Data using id & profile type
      parameters:
        - name : id
          in: query
          description: data id
          required: true
          schema:
            type: integer
            format: int64
        - name: type
          in: query
          description: type of the profiles
          schema: 
            type: string
            enum: 
              - insta
              - facebook
              - linkedin
              - glassdoor
            default: linkedin
      responses: 
        '200': 
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  id: 
                    type: integer
                  type: 
                    type: string
                  followers:
                    type: integer
                  follows:
                    type: integer
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '500':
          description: 'Server Error'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                  error: 
                    type: string
  /add/profile:
    post:
      tags: 
        - Profiles
      summary: Add profile
      consumes: 
        - application/json
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: 
                  type: integer
                name: 
                  type: string
                followers: 
                  type: integer
                follows: 
                  type: integer
      responses: 
        '201':
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  isSuccess: 
                    type: boolean
        '500':
          description: Server error
          content: 
            application/json:
              schema:
                type: object
                properties:
                  isSuccess: 
                    type: boolean
                  error:
                    type: string
  
                

                
                